adding of the module mode_invoicing;
This module allows you to define your preferencies for analytic accounts.

For example, you can say 
* what is the defaut function of a specific user on a given account, 
* what is the default invoicing rate for a specific user on a given account,

This is mostly used when a user encode his timesheet: the values are retrieved and the fields are auto-filled... but the possibility to change these values is still 
available.

Obviously if no data has been recorded for the current account, the default values are given as usual so that this module is perfectly compatible with older configurations.