Account/sequence: try to optimize call wrt. fiscalyear condition.
Note that most objects may not pass the fiscal year into the context.
Note 2: this patch /breaks/ the selection and will return undefined
results when two sequences both with the correct f.y. and without one
exist. In such a case, the administrator will have to put a higher
weight on the right one.