http-well-known: Let it expand, fix body of PROPFIND request
Since this class will be used as a class name (not an object instance),
we need a per-class way to redefine the redirects. Need also to redirect
whole branches rather than path nodes (eg. /principals/... to
/webdav/db/principals/...) .

Finally, solved the problem that do_PROPFIND would not consume the body
of the requests, thus mis-aligning the http layer and breaking the next
request.