webdav: ';' in name not handled correctly [Bug 704444]
Reported by: Dr. Ferdinand Gassauer

Locations that contain a semicolon are split by the urlparse.urlparse(),
thus setting their 'path' and 'params' return fields (rather than just
the 'path'). In our syntax, there is no params, so the part after the
semicolon needs to be appended back to the patch.