[FIX] share: avoid creating ir.rule when o2m has a reverse relationship that is not a real m2o
This case happens rarely, but for example mailgate.message has
an integer res_id column, and some models (eg crm.lead) have
o2m columns that use it as a pseudo-m2o, with the help of a
filter (domain) on the model.
Share wizard will not create ir.rules that use these kind of
pseudo-m2o fields are real m2o.