[FIX] all: use newly introduced fields.date.context_today
Following the introduction of fields.date.context_today,
correct the main cases where we need to use the user's
timezone instead of the default UTC one.
There are probably many minor cases where this may be
useful, but those cases should be easy to fix now
that there is a common way to fix them.