[FIX] project: project preference field must be editable
The user preferences screen is kind of a hack
because it lets the user modify fields on a
model that is normally read-only for her (res.users).
This is accomplished by a custom write() method
on res.users combined with a view on which some
fields are forced to be read-write using a
trick: @readonly=0.