[FIX] account, duplicate of account.tax: better fix for problem fixed in revision 8807.
fixed the duplicate of tax (wasn't working because of a unique constraint on name field). The previous fix was working only for taxes that doesn't have any child. The reason is that copy() is never called for one2many fields, only copy_data() is called (and give the same results for other fields)