[FIX] mail: message_process: do not process incoming emails with a message_id already existing in database.
Before trying to find possible routes, check that the incoming email's
message_id is not already present in mail.message table.
If it is the case, return False.

Parsing of the message has been moved before routing, to avoid looking
for routes for emails we will discard.

Tests have been added and updated.