[REF] report_webkit: unlink/create/write are wrong.
Their changes are not propagated to other processes in a
multi-process setting.

It means that now calling these on auto=False webkit reports
will not have the desired effect. But really they can be updated
to use the new XML parser attribute, and the create/write/unlink
features should be on the base class, not in this module.