[FIX] res.partner form view: inherited view now hides `use_parent_address` on top of `type` + fix embedded form view too
The `use_parent_address` flag is meant for advanced use
in companies that want to make sure their contact/company
addresses are always in sync, at the possible cost of some
usability loss when editing existing addresses (need to
open the company form view, contact fields are read-only).

The embedded mini-form view of the contacts also need to
be inherited to make it match the main view.