[FIX] sale_stock: when creating invoices after delivery, use invoicing contact from SO + update test (reported by Joel Grand-Guillaume while reviewing bug 1160365)
The existing tests were modified to explicitly cover
the case where the "invoice partner" is set on the
Sales Order. This minimal change was done with a
!python block because !record triggers the on_change
calls on the SO and cause undesired changes.
This could be improved by creating a new customer
with a real invoicing contact and creating a new
SO.