[FIX] account: split the actions to open analytic accounts, as context is not compatible
The same action was erroneously used in two incompatible
cases: once for openining the list of analytic accounts
from a partner form view (requires a special context
with a client-side `active_id` variable) and the second
one for the main Analytic Account menu item (which
cannot have `active_id` in the context - it would crash).

The fix includes setting an empty ({}) context explicitly
on the original action to be sure the incorrect context
is removed upon update.


The error was introduced at rev.8685
revid:qdp-launchpad@openerp.com-20130418125951-8p0tfexd9jj8l75b