[FIX] necessary widget implementation
``inline`` is now automatically inferred on a per-widget basis if not
specified.

``inline`` true|false tells the widget whether it should work on
inlines (e.g. span) or blocks (e.g. p). Specifying the wrong one blows
up because the wrapper is incorrect (inline==false on an inline
element will wrap it in a div and add linefeeds, inline==true on a
block element will wrap it in a span which will automatically split up
in multiple elements leading to downcast blowing up because the wrong
wrapper will be selected).

Since last time, inline inference was added: if *unspecified* widget
will check what type of object it's called on (using
``CKEDITOR.dtd.$inline``) and generate the correct wrapper for that
type.

=> fix misbehavior by removing explicit inline spec.

Also remove unused init and data methods.