[FIX] don't use mutation events if unnecessary
Mutation events are synchronous and one is triggered per observed
alteration (they're not batched). When doing lots of modifications to
the DOM tree, they will add a lot of overhead as (in this case) the
event handler will run for *each* added node and it will run during
(and slow down) node insertion rather than afterwards.