[REV] crm: reverted most changes about stages for this branch.
We keep fold, because it has a meaning in kanban view, and is kind of
'drop' stage, for won / lost / cancel / dead.

New stage is now the stage with the lower sequence, not necessarily the
stage with sequence 1. Note that this could break the _track because it
checks for stages with sequence <= 1; however with OpenERP values it
works fine. In some specific stage configuration, some subtypes could
not be tracked precisely.