[FIX] db handling being a pain in the ass
On runbot the base url dispatches into a unknown void of no db being
selected and the CMS thus 404s. For human beings there are additional
links to special subdomains setting the right db in the right place
(or maybe the right dbfilter, whatever) but that doesn't work right
for the crawler test suite.

Instead go through /web/ with a forced db selection, this *should* set
the db correctly in the user's session even if he's anonymous and thus
allows us to browse said db.

No idea if this fevered idea is actually going to work.