[IMP] editability criteria for links and images
if a link or image is an editing host (has @contenteditable=true), its
body is editable *but* its attributes should not be considered
editable: if the attributes come from a t-att-*, @contenteditable will
be set on the node itself.

such elements should be ignored by the link & image dialog handling
thing, as editing their attributes has undesirable effects.