[FIX] odd behavior in cke widgets
in "block-level" widgets (divs), when hitting [return] instead of
adding a newline in the widget a new bit of widget may be added. This
is caused by an incompability between widgets and the enterKey plugin.

Remove enterKey, behavior on [Return] is slightly worse (always
inserts br) but widgets don't get broken.

See https://dev.ckeditor.com/ticket/11326 for issue in upstream.