[FIX] purchase: on po confirm, avoid to write location_id on the chained stock move if the move is done
On purchase order confirmation, if the order_line comes from a procurement order created by a sale order,
it tried to write on the location_id, whatever the move state was, even done.

Or, sometimes, the stock move associated to the procurement order (from the so) was already done, for example because we forced the availability of the stock move.
It makes sense to not write the location_id if the move is already done, because the location_id is already good.