[FIX] mail: for mass mailing, replies in document False by default
Post a copy in the document is False by default, and it does not make sense to replies in the document if the copy is not itself in the document.
Therefore, if post a copy in the document is false by default, replies in the document should be false too.

Moreover, reply_to is required in the form view if mass mailing mode and post a copy is false OR replies in the document is false
But, to avoid updating the view, in the send_mail python function, if no reply to is set and post a copy or replies is false, we set the from email as the reply_to email