[FIX] point_of_sale: validate and closing pos sessoin handle multi company
When the user who validated and closed the pos session was not the same user who created the session, and if this user was not in the same company, it wasnt possible to validate and close the pos session.

This explanation if also valid if the user who created the pos session changed of company in his preferences between the creation and the validation.