[FIX] code cleanup (addon mrp)
change the way bom_count, bom_strct and mo_count are computed, removes the try/except/pass statements, improves the logic.

Also, for some reason, bom_strct is the same as bom_count, but I don't really know what it is supposed to be, and the Bom Structure and Bill of materials buttons leads to views with the same number of items.