[FIX] product: when duplicating a product, keep the language in the context
At the time of the context_wo_lang patch (7.0 revision 6577), the orm did not keep the language in copy_data, this patch intended to be more consistent.
Since server revision 5146 7.0, the new behaviour is to use the translated version in copy_data. Removign this change will be more consistent with the orm.

The expected behaviour is now the following:
In user lang: translated product name + translated '(copy)'
In other lang: same as original product